+++ toolkit_test.go
@@ -7,16 +7,18 @@ import (
 )
 
 func TestToolkit_ConfigInitOption(t *testing.T) {
-	got := ConfigInitOption("./test", "{}")
-	require.True(t, got)
+	err := ConfigInitOption("./test", "{}")
+	require.NoError(t, err)
 }
 
 func TestToolkit_FindAdbDevices(t *testing.T) {
-	adbDevices := FindAdbDevices()
+	adbDevices, err := FindAdbDevices()
+	require.NoError(t, err)
 	require.NotNil(t, adbDevices)
 }
 
 func TestToolkit_FindDesktopWindows(t *testing.T) {
-	desktopWindows := FindDesktopWindows()
+	desktopWindows, err := FindDesktopWindows()
+	require.NoError(t, err)
 	require.NotNil(t, desktopWindows)
 }